requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.8
Django==4.2
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-filter==23.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
